refactor: putting types with their associated code
Also simplifying exports by using wildcards (*) where possible.
