+++ index.ts
@@ -1,148 +1,7 @@
-export {
-  APP_ANALYTICS_INITIALIZED,
-  APP_AUTH_INITIALIZED,
-  APP_CONFIG_INITIALIZED,
-  APP_I18N_INITIALIZED,
-  APP_INIT_ERROR,
-  APP_LOGGING_INITIALIZED,
-  APP_PUBSUB_INITIALIZED,
-  APP_READY,
-  APP_TOPIC,
-  AppContext,
-  AppProvider,
-  auth,
-  AUTHENTICATED_USER_CHANGED,
-  AUTHENTICATED_USER_TOPIC,
-  AuthenticatedPageRoute,
-  AxiosJwtAuthService,
-  camelCaseObject,
-  clearAllSubscriptions,
-  CONFIG_CHANGED,
-  CONFIG_TOPIC,
-  configureAnalytics,
-  configureAuth,
-  configureI18n,
-  configureLogging,
-  convertKeyNames,
-  createIntl,
-  defineMessages,
-  Divider,
-  ensureAuthenticatedUser,
-  ErrorBoundary,
-  ErrorPage,
-  fetchAuthenticatedUser,
-  FormattedDate,
-  FormattedMessage,
-  FormattedNumber,
-  FormattedPlural,
-  FormattedRelativeTime,
-  FormattedTime,
-  getAnalyticsService,
-  getAuthenticatedHttpClient,
-  getAuthenticatedUser,
-  getAuthService,
-  getBasename,
-  getConfig,
-  getHistory,
-  getHttpClient,
-  getLocale,
-  getLocalizedLanguageName,
-  getLoggingService,
-  getLoginRedirectUrl,
-  getLogoutRedirectUrl,
-  getMessages,
-  getPath,
-  getPrimaryLanguageSubtag,
-  getQueryParameters,
-  getSupportedLanguageList,
-  handleRtl,
-  hydrateAuthenticatedUser,
-  identifyAnonymousUser,
-  identifyAuthenticatedUser,
-  initError,
-  initialize,
-  initializeMockApp,
-  injectIntl,
-  IntlProvider,
-  intlShape,
-  isRtl,
-  LOCALE_CHANGED,
-  LOCALE_TOPIC,
-  logError,
-  logInfo,
-  LoginRedirect,
-  mergeConfig,
-  mergeMessages,
-  MockAnalyticsService,
-  MockAuthService,
-  MockLoggingService,
-  mockMessages,
-  modifyObjectKeys,
-  NewRelicLoggingService,
-  PageWrap,
-  parseURL,
-  Plugin,
-  PluginSlot,
-  publish,
-  redirectToLogin,
-  redirectToLogout,
-  resetAnalyticsService,
-  resetLoggingService,
-  SegmentAnalyticsService,
-  sendPageEvent,
-  sendTrackEvent,
-  sendTrackingLogEvent,
-  setAuthenticatedUser,
-  setConfig,
-  snakeCaseObject,
-  subscribe,
-  unsubscribe,
-  updateLocale,
-  useAppEvent,
-  useAuthenticatedUser,
-  useConfig,
-  useIntl
-} from './runtime';
+export * from './runtime';
 
-export type {
-  App,
-  BaseWidgetOperation,
-  ComponentOperation,
-  CourseInfo,
-  ElementOperation,
-  ExternalRoute,
-  FederatedApp,
-  FederatedOperation,
-  IFrameOperation,
-  LocalizedMessages,
-  MenuItemName,
-  LayoutOptionsOperation,
-  ProjectModuleConfig,
-  ProjectSiteConfig,
-  Remote,
-  ReplaceLayoutOperation,
-  SlotCondition,
-  SlotOperation,
-  User,
-  WidgetOperation
-} from './types';
+export type * from './types';
 
-export {
-  EnvironmentTypes,
-  LayoutOperationTypes,
-  PluginOperationTypes,
-  PluginTypes,
-  WidgetOperationTypes
-} from './types';
+export * from './types';
 
-export {
-  defaultFooterConfig,
-  defaultHeaderConfig,
-  DefaultLayout,
-  DefaultMain,
-  defaultShellConfig,
-  Footer,
-  Header,
-  LinkMenuItem,
-  NavDropdownMenuSlot
-} from './shell';
+export * from './shell';
