+++ runtime/slots/index.ts
@@ -1,0 +1,10 @@
+export type * from './types';
+export {
+  LayoutOperationTypes,
+  WidgetOperationTypes
+} from './types';
+
+export { default as DefaultSlotLayout } from './DefaultSlotLayout';
+export { default as FederatedWidget } from './FederatedWidget';
+export { default as Slot } from './Slot';
+export { default as SlotContext } from './SlotContext';
